requirements.txt
streamlit
openai
langchain
langchain-openai
langchain-community
pypdf2
pillow
python-magic-bin
pytesseract
pandas